Create a new User Follow repository that will handle follow/following stuff.
