+++ app/src/main/java/com/diargegaj/recipesharing/data/repository/UserRepositoryImpl.kt
@@ -24,83 +24,6 @@ class UserRepositoryImpl @Inject constructor(
     private val fireStore: FirebaseFirestore
 ) : UserRepository {
 
-
-    override fun isUserFollowing(
-        loggedInUserId: String,
-        otherUserId: String
-    ): Flow<Resource<Boolean>> = flow<Resource<Boolean>> {
-        val docSnapshot = fireStore.collection("following")
-            .document(loggedInUserId)
-            .collection("userFollowing")
-            .document(otherUserId)
-            .get()
-            .await()
-
-        emit(Resource.Success(docSnapshot.exists()))
-    }.catch { throwable ->
-        emit(
-            Resource.Error(Exception(throwable))
-        )
-    }
-
-    override suspend fun followUser(loggedInUserId: String, otherUserId: String): Resource<Unit> {
-        return safeCall {
-            val batch = fireStore.batch()
-
-            val loggedInUserRef =
-                fireStore.collection(DBCollection.User.collectionName).document(loggedInUserId)
-            batch.update(loggedInUserRef, "followingCount", FieldValue.increment(1))
-
-            val otherUserRef =
-                fireStore.collection(DBCollection.User.collectionName).document(otherUserId)
-            batch.update(otherUserRef, "followersCount", FieldValue.increment(1))
-
-            val followingRef =
-                fireStore.collection(DBCollection.Following.collectionName).document(loggedInUserId)
-                    .collection(DBCollection.UserFollowing.collectionName).document(otherUserId)
-            batch.set(followingRef, mapOf("followingUserId" to otherUserId))
-
-            val followersRef =
-                fireStore.collection(DBCollection.Followers.collectionName).document(otherUserId)
-                    .collection(DBCollection.UserFollowers.collectionName)
-                    .document(loggedInUserId)
-            batch.set(followersRef, mapOf("followerUserId" to loggedInUserId))
-
-            batch.commit().await()
-
-            Resource.Success(Unit)
-        }
-    }
-
-    override suspend fun unfollowUser(loggedInUserId: String, otherUserId: String): Resource<Unit> {
-        return safeCall {
-            val batch = fireStore.batch()
-
-            val loggedInUserRef =
-                fireStore.collection(DBCollection.User.collectionName).document(loggedInUserId)
-            batch.update(loggedInUserRef, "followingCount", FieldValue.increment(-1))
-
-            val otherUserRef =
-                fireStore.collection(DBCollection.User.collectionName).document(otherUserId)
-            batch.update(otherUserRef, "followersCount", FieldValue.increment(-1))
-
-            val followingRef =
-                fireStore.collection(DBCollection.Following.collectionName).document(loggedInUserId)
-                    .collection(DBCollection.UserFollowing.collectionName).document(otherUserId)
-            batch.delete(followingRef)
-
-            val followersRef =
-                fireStore.collection(DBCollection.Followers.collectionName).document(otherUserId)
-                    .collection(DBCollection.UserFollowers.collectionName)
-                    .document(loggedInUserId)
-            batch.delete(followersRef)
-
-            batch.commit().await()
-
-            Resource.Success(Unit)
-        }
-    }
-
     override suspend fun getFollowersForUser(
         userId: String,
         currentUserId: String
@@ -125,16 +48,16 @@ class UserRepositoryImpl @Inject constructor(
                     userDto?.userUUID = followerId
                     userDto?.let {
                         val userModel = it.mapToDomain()
-                        val isFollowing = when (
-                            val result = isUserFollowing(
-                                currentUserId,
-                                followerId
-                            ).first()
-                        ) {
-                            is Resource.Success -> result.data
-                            else -> false
-                        }
-                        userModel.isFollowedByCurrentUser = isFollowing
+//                        val isFollowing = when (
+//                            val result = isUserFollowing(
+//                                currentUserId,
+//                                followerId
+//                            ).first()
+//                        ) {
+//                            is Resource.Success -> result.data
+//                            else -> false
+//                        }
+//                        userModel.isFollowedByCurrentUser = isFollowing
                         followers.add(userModel)
                     }
                 }
@@ -167,17 +90,17 @@ class UserRepositoryImpl @Inject constructor(
                     userDto?.userUUID = followingId
                     userDto?.let {
                         val userModel = it.mapToDomain()
-                        val isFollowing = when (
-                            val result = isUserFollowing(
-                                currentUserId,
-                                followingId
-                            ).first()
-                        ) {
-                            is Resource.Success -> result.data
-                            else -> false
-                        }
-                        userModel.isFollowedByCurrentUser = isFollowing
-                        followingUsers.add(userModel)
+//                        val isFollowing = when (
+//                            val result = isUserFollowing(
+//                                currentUserId,
+//                                followingId
+//                            ).first()
+//                        ) {
+//                            is Resource.Success -> result.data
+//                            else -> false
+//                        }
+//                        userModel.isFollowedByCurrentUser = isFollowing
+//                        followingUsers.add(userModel)
                     }
                 }
                 Resource.Success(followingUsers)
