+++ app/src/main/java/com/diargegaj/recipesharing/di/module/RepositoryModule.kt
@@ -1,14 +1,16 @@
 package com.diargegaj.recipesharing.di.module
 
 import com.diargegaj.recipesharing.data.repository.FirebaseImageUploadRepositoryImpl
-import com.diargegaj.recipesharing.data.repository.recipe.RecipeRepositoryImpl
 import com.diargegaj.recipesharing.data.repository.UserRepositoryImpl
+import com.diargegaj.recipesharing.data.repository.recipe.RecipeRepositoryImpl
 import com.diargegaj.recipesharing.data.repository.userAuth.UserAuthRepositoryImpl
+import com.diargegaj.recipesharing.data.repository.userFollow.UserFollowRepositoryImpl
 import com.diargegaj.recipesharing.data.repository.userProfile.UserProfileRepositoryImpl
 import com.diargegaj.recipesharing.domain.repository.ImageUploadRepository
 import com.diargegaj.recipesharing.domain.repository.RecipeRepository
 import com.diargegaj.recipesharing.domain.repository.UserRepository
 import com.diargegaj.recipesharing.domain.repository.userAuth.UserAuthRepository
+import com.diargegaj.recipesharing.domain.repository.userFollow.UserFollowRepository
 import com.diargegaj.recipesharing.domain.repository.userProfile.UserProfileRepository
 import dagger.Binds
 import dagger.Module
@@ -43,5 +45,10 @@ abstract class RepositoryModule {
     abstract fun provideUserProfileRepository(
         userProfileRepositoryImpl: UserProfileRepositoryImpl
     ): UserProfileRepository
+
+    @Binds
+    abstract fun provideUserFollowRepository(
+        userFollowRepositoryImpl: UserFollowRepositoryImpl
+    ): UserFollowRepository
 
 }
