+++ app/src/main/java/com/diargegaj/recipesharing/domain/repository/userFollow/UserFollowRepository.kt
@@ -1,0 +1,14 @@
+package com.diargegaj.recipesharing.domain.repository.userFollow
+
+import com.diargegaj.recipesharing.domain.utils.Resource
+import kotlinx.coroutines.flow.Flow
+
+interface UserFollowRepository {
+
+    fun isUserFollowing(loggedInUserId: String, otherUserId: String): Flow<Resource<Boolean>>
+
+    suspend fun followUser(loggedInUserId: String, otherUserId: String): Resource<Unit>
+
+    suspend fun unfollowUser(loggedInUserId: String, otherUserId: String): Resource<Unit>
+
+}
